+++ packages/blocks/core/js/wordpress/buttons/test/functionality.blocks-1.e2e.cy.js
@@ -37,6 +37,7 @@ describe('Buttons Block', () => {
 
 		// Block supported is active
 		cy.get('.blockera-extension-block-card').should('be.visible');
+
 		//
 		// 1. Edit Block
 		//
@@ -96,7 +97,19 @@ describe('Buttons Block', () => {
 			});
 
 		//
-		// 2. Assert inner blocks selectors in front end
+		// 2. Check settings tab
+		//
+		cy.getByDataTest('settings-tab').click();
+
+		// layout settings should be hidden
+		cy.get('.block-editor-block-inspector').within(() => {
+			cy.get('.components-panel__body-title')
+				.contains('Layout')
+				.should('not.be.visible');
+		});
+
+		//
+		// 3. Assert inner blocks selectors in front end
 		//
 		savePage();
 		redirectToFrontPage();