+++ packages/blocks/core/CHANGELOG.md
@@ -81,6 +81,7 @@
 - `Audio` block: support improved.
 - `Avatar` block: support improved.
 - `Button` block: support improved.
+- `Buttons` block: support improved.
 
 
 ### Automated Tests
@@ -99,6 +100,7 @@
 - Updated E2E tests to check support of `Audio` block.
 - Updated E2E tests to check support of `Avatar` block.
 - Updated E2E tests to check support of `Button` block.
+- Updated E2E tests to check support of `Buttons` block.
 
 
 ## 1.3.0 (2025-04-27)